# theLinkWeb/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.1.0"
  title: theLink API
# during dev, should point to your local machine
host: localhost:3050
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - application/x-www-form-urlencoded
# format of the responses to the client (Accepts)
produces:
  - application/json
  - text/html
...
paths:
...
  /api/pivot/info/{id}:
    x-swagger-router-controller: pivot
    get:
      tags:
        - Pivot
      operationId: getInfo
      description: Gets pivot info for a given node
      parameters:
        - name: id
          in: path
          description: The ID of the node. (Can by any type of node)
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GenericResponse"  
  /api/pivot/{id}:
    x-swagger-router-controller: pivot
    get:
      tags:
        - Pivot
      operationId: pivotAll
      description: Pivots over all relationships a given node has.
      parameters:
        - name: id
          in: path
          description: The ID of the node. (Can by any type of node)
          required: true
          type: string
        - $ref: "#/parameters/group"
        - $ref: "#/parameters/groupcutoff"
        - $ref: "#/parameters/firstLetter"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GenericResponse"  
  /api/pivot/bytype/{id}/{nodetype}:
    x-swagger-router-controller: pivot
    get:
      tags:
        - Pivot
      operationId: pivotByType
      description: Pivots over a node, returning all nodes of a certain type with a given relationship to the supplied node.
      parameters:
        - name: id
          in: path
          description: The ID of the node. (Can by any type of node)
          required: true
          type: string
        - name: nodetype
          in: path
          description: The type of node we are pivoting over. Only nodes of this type will be returned.
          required: true
          type: string
        - $ref: "#/parameters/group"
        - $ref: "#/parameters/groupcutoff"
        - $ref: "#/parameters/firstLetter"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GenericResponse"  
  /api/pivot/byrel/{id}/{reltype}:
    x-swagger-router-controller: pivot
    get:
      tags:
        - Pivot
      operationId: pivotByRel
      description: Pivots over a specific relationship on a node, returning all nodes with that relationship to this node. Relationships work in either direction.
      parameters:
        - name: id
          in: path
          description: The ID of the node. (Can by any type of node)
          required: true
          type: string
        - name: nodetype
          in: query
          description: The type of node we are pivoting over. Only nodes of this type will be returned.
          required: false
          type: string
        - name: reltype
          in: path
          description: The type of relationship we are pivoting over.
          required: true
          type: string
        - $ref: "#/parameters/group"
        - $ref: "#/parameters/groupcutoff"
        - $ref: "#/parameters/firstLetter"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GenericResponse"
...
  /api/questions/dashboard:
    x-swagger-router-controller: questions
    get:
      operationId: dashboard
      tags:
        - Questions
      parameters:
        - name: label
          in: query
          description: The label from the dashboard graph that the question is about
          type: string
          required: true
        - name: questionstr
          in: query
          description: The string corresponding to the question we want to ask
          type: string
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/GenericResponse"            
...
parameters:
...
  firstLetter:
    name: firstletter
    in: query
    description: Returns items whose name starts with this letter
    type: string
...
  groupcutoff:
    name: groupcutoff
    in: query
    description: If grouping enabled, only group if we have more than this number of results
    type: integer
  group:
    name: group
    in: query
    description: Return items grouped by 'alphabet', 'date', or 'location'
    type: string
...
definitions:
  GenericResponse:
    type: object
    #required:
    #  - data
    #properties:
    #  data:
    #    type : array
...